+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "NGR basic radar".
 - Tests for "1000000000D world".
 - Tests for "1D bush fire".
+- Tests for "Rectangle partition".
 
 ## [1.4.0] - 2022-03-03
 ### Added